test(desktop): split cached fetch specs
