+++ scripts/check-file-sizes.mjs
@@ -68,7 +68,6 @@ const ALLOWLIST = new Set([
 
   // ── Test files (allow-listed; not a Phase 4 priority) ──────────
   "apps/desktop/src/__tests__/components/timeline/AgentTreeView.test.ts",
-  "apps/desktop/src/__tests__/composables/useCachedFetch.test.ts",
   "apps/desktop/src/__tests__/views/analytics-views.test.ts",
 ]);
